pyproject.toml
[tool.poetry]
name = "teste"
version = "0.1.0"
description = ""
authors = ["plinio-jr <pliniojunior5@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.3"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.13.0"
djangorestframework-simplejwt = "^5.2.2"
drf-spectacular = "^0.24.2"
python-magic = "^0.4.27"
Pillow = "^9.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
